// frontend/src/api/models/handlerWarehouseResponse.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */
import type { HandlerWarehouseResponseStatus } from './handlerWarehouseResponseStatus'
import type { HandlerWarehouseResponseType } from './handlerWarehouseResponseType'

/**
 * Warehouse details returned by the API
 */
export interface HandlerWarehouseResponse {
  address?: string
  attributes?: string
  city?: string
  code?: string
  country?: string
  created_at?: string
  email?: string
  full_address?: string
  id?: string
  is_default?: boolean
  manager_name?: string
  name?: string
  notes?: string
  phone?: string
  postal_code?: string
  province?: string
  short_name?: string
  sort_order?: number
  status?: HandlerWarehouseResponseStatus
  tenant_id?: string
  type?: HandlerWarehouseResponseType
  updated_at?: string
  version?: number
}